Create package.json
The error message from the tests failing indicates that npm is unable to find the package.json file in the specified directory, which is causing the ENOENT error. This file is typically required for npm to run properly, so this error is preventing the test from running. To fix the error, you will need to ensure that the package.json file is present in the specified directory.

I copied the code from Matt's playwright framework in the link below and added it to the package.json file here. https://github.com/GuardianUI/PlaywrightFramework/blob/master/package.json